### 2024-05-26 #6791 @editorialbot
[PRE REVIEW]: Equilipy: a python package for calculating phase equilibria
**Submitting author:** @SunyongKwon (<a href="http://orcid.org/0000-0002-9212-8823">Sun Yong Kwon</a>)
**Repository:** https://github.com/ORNL/Equilipy
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.3
