### 2024-05-29 #6791 @phibeck
Hi @SunyongKwon, thanks for your submission. I'll be looking for reviewers next, thank you for the suggestions.
In the meantime:
I noticed that the paper is on the longer side (we aim for <= 1000 words, but not strict), so you could think about moving the features section into the repository documentation directly, assuming it's more likely for future users to look for updated documentation there instead of the JOSS paper.
