### 2024-05-29 #6791 @SunyongKwon
Hi @phibeck,

Thank you for the suggestion. I am actually aware that my manuscript exceeds the journal's word limit. However, I would like to retain the feature section in the manuscript. The main goal of this section is to compare the performance of this package with other commonly-used (commercial) software in my research area, which is one of the review criteria of JOSS. Please let me know if it is still okay to keep the length of the manuscript.

Thank you again for the consideration!
