### 2024-06-11 #6791 @phibeck
> Hi @phibeck,
>
> It's been two weeks I haven't heard from you. Do you mind let me know if the manuscript is under review process?
>
> Best regards, Sunyong Kwon

Thank you for your patience. Since you expressed that you would like to keep the sections in the manuscript, and indeed the features section focuses on performance comparison, let's keep it as is.
