### 2024-06-11 #6875 @editorialbot
[REVIEW]: Equilipy: a python package for calculating phase equilibria
**Submitting author:** @SunyongKwon (<a href="http://orcid.org/0000-0002-9212-8823">Sun Yong Kwon</a>)
**Repository:** https://github.com/ORNL/Equilipy
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.5
