### 2024-06-28 #6875 @mefuller
I am also aiming to complete this in the coming two weeks as work has been quite hectic until now.
*Mark E. Fuller, Ph.D.*
***@***.****
***@***.****
***@***.***:fedora.im*
***@***.***:one.ems.host*
*https://mefuller.dev*
*PGP Fingerprint: 73F1 A30C BDF4 DB4B C75F FD0F D599 E76C FFCA BF60*

Jun 28, 2024 17:46:00 Pierre Lanari ***@***.***>:

>
> @phibeck[https://github.com/phibeck] @SunyongKwon[https://github.com/SunyongKwon] I am very busy right now, but do not worry, I will finish my review within the next 2 weeks. I will create the checklist in time.
>
