### 2024-07-09 #6875 @lanari
@SunyongKwon please accept my apologies for the delay in reviewing your work.

@phibeck I have finished my review and here are my comments:
- This is a good package that will be useful to the community as this area of research is unfortunately still dominated by commercial software.
- The repository contains all the necessary files and the examples provided are easy to follow and reproduce. I have added a comment on an issue created by @mefuller in the repository.
- I agree with @mefuller that a simple automated test with a performance report would be a great addition. I could run all the examples, but I don't know how the performance on my machine compares to other setups.
- The manuscript is well written and easy to follow. A few minor suggestions for @SunyongKwon
    - I suggest using italics for variables (_N_, _T_ and _P_ in line 46 and in Figure 2; _G_, _H_, _S_, _Cp_ in lines 62-63). It is common practice to report variables in italics in scientific papers to distinguish them from abbreviations, labels or element names. For example, P is phosphorus, whereas _P_ is pressure.
    - In Table 1, I think it would be interesting to indicate which software solutions can also use the thermodynamic data used in this study. Holland & Powell's THERMOCALC does not perform Gibbs Free Energy minimisations (see discussion in Lanari, P. & Duesterhoeft, E. (2019). Modelling metamorphic rocks using equilibrium thermodynamics and internally consistent databases: past achievements, problems and perspectives. Journal of Petrology, 60, 19-56).
    - Finally, it might be interesting to add a few sentences about how the CALPHAD approach differs from other approaches used.

I hope that some of my comments will be helpful.
