+++ setup.py
@@ -73,7 +73,7 @@ class f2py_Build(build_ext):
             
 setup(
     name='equilipy',
-    version='0.1.3',
+    version='0.1.4',
     author='Sunyong Kwon',
     author_email='sunyong.kwon@mail.mcgill.ca',
     license='BSD-3',
