### 2024-07-15 #6875 @SunyongKwon
Hi @lanari

Thank you so much for the thorough review. I did not know that Holland & Powell's THERMOCALC does not ensure the stability of the equilibrated assemblage for a specific bulk composition.
I will revise documentation and manuscript accordingly and keep you updated.

Thank you so much again for your insightful feedback.
