+++ paper/paper.bib
@@ -263,3 +263,15 @@
   publisher={Springer},
   doi={10.1007/s11661-017-4059-0}
 }
+
+@article{Lanari:2019,
+  title={Modeling metamorphic rocks using equilibrium thermodynamics and internally consistent databases: Past achievements, problems and perspectives},
+  author={Lanari, Pierre and Duesterhoeft, Erik},
+  journal={Journal of Petrology},
+  volume={60},
+  number={1},
+  pages={19--56},
+  year={2019},
+  publisher={Oxford University Press}
+    doi={10.1093/petrology/egy105}
+}
