### 2024-07-16 e37f952
Add the description between CALPHAD and Geology models
