+++ paper/paper.md
@@ -87,9 +87,11 @@ Exploring Scheil-Gulliver solidification offers a robust framework for assessing
 Equilipy leverages parallel computing to calculate multiple input conditions simultaneously through its `equilib_batch()` function. Since computations between input conditions are fully independent, they are executed through embarrassingly parallel runs. Note that parallel computing is not applied to compute a single input condition. By default, Equilipy uses all available processors on a single computing node, though users have the option to specify a different number of processors. In multi-node environments, users are required to configure the number of nodes and processors using a Message Passing Interface (MPI) such as `OpenMPI` or `MPICH`. Equilipy employs `mpi4py` for interfacing with MPI programs.
 
 # Related programs
-**Table 1** presents a compilation of both commercial and non-commercial programs similar to Equilipy, designed for computing multicomponent-multiphase equilibria. Broadly, these programs split into two categories: those employing the CALPHAD approach, and those designed for geological applications. The primary distinction between these two categories lies in their respective free energy model descriptions. ADD DESCRIPTION ON CALPHAD(TBD)  These variations in free energy descriptions are provided in literature for the CALPHAD approach [@CALPHAD:1998] and for geological applications [@Holland:2011].
+**Table 1** presents a compilation of both commercial and non-commercial programs similar to Equilipy, designed for computing multicomponent-multiphase equilibria. Broadly, these programs split into two categories: those employing the CALPHAD approach, and those designed for geological applications. While both approaches share the fundamental principle of using Gibbs energies as the building blocks of thermochemical properties, assessed from experimental data, their primary distinction lies in their free energy model descriptions.
 
-Note that there is no general consensus on thermochemical database format. Each software often has its own thermochemical database format. The thermochemical database used in Equilipy `.dat` is only compatible with the commercial software FactSage [@FactSage:2016].
+For geological applications, the gibbs energies of non-ideal solution phases are described by Margules parameters as a function of composition, temperature and pressure. In contrast, the CALPHAD approach focuses on structure-based Gibbs energy descriptions at 1 atm pressure using solution models such as the compound energy formalism (CEF) for long-range ordering and the modified quasichemical models for short-range ordering. These variations in free energy descriptions are well-documented in the literature for the CALPHAD approach [@CALPHAD:1998] and for geological applications [@Holland:2011].
+
+Note that there is no general consensus on thermochemical database format. Each software often has its own thermochemical database format. The thermochemical database used in Equilipy `.dat` is compatable in FactSage [@FactSage:2016], PyCalphad [@PyCALPHAD:2017], and Thermochimica [@Thermochimica:2013].
 
 **Table 1** List of commercial and non-commercial programs for multicomponent-multiphase equilibrium calculations.\label{Table1}
 
