### 2024-07-17 1dcfd2e
debug paper.bib
