+++ paper/paper.bib
@@ -272,6 +272,6 @@
   number={1},
   pages={19--56},
   year={2019},
-  publisher={Oxford University Press}
-    doi={10.1093/petrology/egy105}
+  publisher={Oxford University Press},
+  doi={10.1093/petrology/egy105}
 }
