### 2024-07-17 4c4c64d
debug paper.md
