+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 # Summary
 The CALPHAD (CALculation of PHAse Diagram) approach [@CALPHAD:1998] provides predictions for thermodynamically stable phases in multicomponent-multiphase materials across a wide range of temperatures. Consequently, the CALPHAD calculations became an essential tool in materials and process design [@Luo:2015]. Such design tasks frequently require navigating a high-dimensional space due to multiple components involved in the system. This increasing complexity demands for high-throughput CALPHAD calculations, especially in the rapidly evolving field of alloy design.
 
-In response to the need, we developed Equilipy an open-source Python package designed for calculating phase equilibria of multicomponent-multiphase systems. Equilipy is specifically tailored for high-throughput CALPHAD calculations, offering parallel computations across multiple processors and nodes with the given NTP input conditions namely elemental compositions (*N*), temperature (*T*), and pressure (*P*). Equilipy utilizes the program structure and Gibbs energy functions from the Fortran-based program, Thermochimica [@Thermochimica:2013], with incorporating a new Gibbs energy minimization algorithm. This algorithm, originally developed by Capitani and Brown in 1987 [@Capitani:1987], has been revised and implemented to enhance the stability and performance of calculations. The Fortran codes are precompiled and interfaced with Python via `F2PY`, ensuring high computation speed. Benchmark tests shown in \autoref{Fig1} demonstrate that Equilipy's computation speed is comparable to those of established commercial software, TC-Python and PanPython. This result highlights its efficiency and potential applications in various scientific and industrial fields.
+In response to the need, we developed Equilipy an open-source Python package designed for calculating phase equilibria of multicomponent-multiphase systems. Equilipy is specifically tailored for high-throughput CALPHAD calculations, offering parallel computations across multiple processors and nodes with the given *NPT* input conditions namely elemental compositions (*N*), pressure (*P*), and temperature (*T*). Equilipy utilizes the program structure and Gibbs energy functions from the Fortran-based program, Thermochimica [@Thermochimica:2013], with incorporating a new Gibbs energy minimization algorithm. This algorithm, originally developed by Capitani and Brown in 1987 [@Capitani:1987], has been revised and implemented to enhance the stability and performance of calculations. The Fortran codes are precompiled and interfaced with Python via `F2PY`, ensuring high computation speed. Benchmark tests shown in \autoref{Fig1} demonstrate that Equilipy's computation speed is comparable to those of established commercial software, TC-Python and PanPython. This result highlights its efficiency and potential applications in various scientific and industrial fields.
 
 ![A parallel benchmark computation on (a) Windows (Intel Core i7-12700, 2.1 GHz) and (b) Linux (Intel Xeon E5-2650, 2.0 GHz) for multiple processors: The computation involves 8,096 equilibrium calculations in a four-component system that has 63 different phases.\label{Fig1}](figures/Fig1.svg){ width=98% }
 
@@ -43,7 +43,7 @@ Commercial thermochemical software packages such as Thermo-Calc [@ThermoCalc:200
 Given these limitations and the pressing need for more effective alloy design processes, developing an open-source program tailored for high-throughput CALPHAD calculations becomes imperative. The envisioned program should not only eliminate the constraints on processor usage, but also ensure high calculation stability. Furthermore, the program should operates on Linux system and support parallel computing to be compatible with HPC environments. An intuitive program architecture is preferred to enhance accessibility for researchers and practitioners in the field.
 
 # Software overview
[…]
