### 2024-07-18 638cdc0
Comment test suite in install.md
