### 2024-07-19 #6875 @SunyongKwon
Dear @phibeck @mefuller @lanari

First of all, thank you for the thorough review. I agree with all the feedback that @mefuller and @lanari provided and revised the manuscript and program accordingly. Please, find the [revised manuscript](https://github.com/ORNL/Equilipy/blob/joss/paper/JOSS_Revised.pdf). I addressed the revision in each issue created in the repository.

Here is the summary of all changes.
1. Add/property document tests and integrate into CI
	-  I added the [CI/CD script](https://github.com/ORNL/Equilipy/blob/main/.github/workflows/build_wheels.yml) in the repository, which utilized GitHub Actions for generating `.whl` files
	- I also added a [Test suite](https://github.com/ORNL/Equilipy/tree/main/test) in the repository and integrated it in the CI/CD workflow.
	- Short documentation regarding [Test suite](https://github.com/ORNL/Equilipy/tree/main/test) is also added in the web document under Install section.
2. Example 3 or (polars) require `fastexcel` package, add this as recommended install (preinstall) under the example of web documentation
	- The description on the dependency of `fastexcel` package was added in `README` and the web documentation under `PreInstall` and `Features and Examples`
3. Add clear documentation of database file format and link for ChemSage.dat
	- I created the database section in the web documentation and link to ChemApp online document for database format.
4. Make sure the versions of softwares are all consistent with manuscript
	- All software versions are bumped up to 0.1.4
5. Suggest using italics for variables (_N_, _T_ and _P_ in line 46 and in Figure 2; _G_, _H_, _S_, _Cp_ in lines 62-63).
	- I revised the font of variables (i.e., N, T, P, G, H, S, and Cp) into Italics in the revised manuscript
6.  In Table 1, I think it would be interesting to indicate which software solutions can also use the thermodynamic data used in this study.
	- Thermochemical database provided in equilipy is only compatible with a commercial software, FactSage and non-commercial packages PyCalphad and Thermochimica. I added the description on the availability of using current database in `Related Program` section of the manuscript
	-  I included footnote about the Holland & Powell's THERMOCALC for it's incapability of searching for thermodynamically stable phase assemblages. I cited the paper @lanari mentioned in the review.
7.  It might be interesting to add a few sentences about how the CALPHAD approach differs from other approaches used.
	- A short description describing the deference between CALPHAD approach and Geological models are provided in `Related Program` of the manuscript

Please let me know if the issues are properly addressed and if you have additional comments. Thank you again for your insightful feedback.
