### 2024-07-26 #6875 @SunyongKwon
Hi @phibeck

Please let me know anytime, if you need any information from me to proceed for the acceptance.

Regarding the missing doi, please note that 'CALPHAD (Calculation of Phase Diagrams): A Comprehensive Guide by N. Saunders and A.P. Miodownik, 1998' does not have doi assigned to it.
