### 2024-07-26 #6875 @phibeck
Hi @SunyongKwon, thanks. A couple of comments:

- Since I cannot find a record of some of your co-authors contribution in your repository, could you please state their contributions here for the record of review?  You can check out the guidance for authorship here:  https://joss.readthedocs.io/en/latest/submitting.html#authorship
- could you please update the title of the archived version to match that of the paper?
- it would be great if you could unify your version tags. It is `0.1.4` in your repo, but `v0.1.4` in the archive. Thank you!
