### 2024-07-26 #6875 @SunyongKwon
Dear @phibeck,

Thank you so much for the clear instruction. I revised the following items according to your suggestion:

- [x] Updated the version from `v0.1.4` to `0.1.4` in Zenodo archive for title, description, and version tag [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.12801779.svg)](https://doi.org/10.5281/zenodo.12801779)
- [x] Updated figure 2 in the [repo](https://github.com/ORNL/Equilipy/tree/joss) since the figure is not clear in the generated pdf.

Here is the description of contribution for my co-authors
**Kwon, Sun Yong**: administrated the project, developed software and algorithms, and wrote the original manuscript
**Thibodeau, Eric**: contributed in literature review and developing the algorithm used in the software
**Alex, Plotkowski**: acquired funding, provided supervision, and editing the manuscript
**Yang, Ying**: Provided thermochemical property database, supervision on algorithm development.
