### 2024-07-26 #6875 @SunyongKwon
Do you mind re-generating the pdf? The figure 2 is being complied correctly now.
Thank you so much!
