### 2024-07-26 #6875 @phibeck
> Do you mind re-generating the pdf? The figure 2 is being complied correctly now. Thank you so much!

No problem! Just want to let you know for the future that a subset of the EditorialBot commands are available to authors and reviewers, including the one that generates the pdf.
