### 2024-07-26 #6875 @phibeck
> Dear @phibeck,
>
> Thank you so much for the clear instruction. I revised the following items according to your suggestion:
>
>     * [x]  Updated the version from `v0.1.4` to `0.1.4` in Zenodo archive for title, description, and version tag [![DOI](https://camo.githubusercontent.com/8cb57fff8fe77c546935e1c4f90767bd9063919e4a0f15583ca77bf51761bba1/68747470733a2f2f7a656e6f646f2e6f72672f62616467652f444f492f31302e353238312f7a656e6f646f2e31323830313737392e737667)](https://doi.org/10.5281/zenodo.12801779)
>
>     * [x]  Updated figure 2 in the [repo](https://github.com/ORNL/Equilipy/tree/joss) since the figure is not clear in the generated pdf.
>
>
> Here is the description of contribution for my co-authors **Kwon, Sun Yong**: administrated the project, developed software and algorithms, and wrote the original manuscript **Thibodeau, Eric**: contributed in literature review and developing the algorithm used in the software **Alex, Plotkowski**: acquired funding, provided supervision, and editing the manuscript **Yang, Ying**: Provided thermochemical property database, supervision on algorithm development.

Thank you very much for clarifying the author contribution and for unifying the version numbers. Sorry if my previous point wasn't sufficiently clear, but could you please make sure that the archive title also matches the title of the paper? Then we're all set.
