### 2024-07-29 #6875 @phibeck
@SunyongKwon here are a few more comments/suggestions for the manuscript. Please have a look when you have a moment.

- line 6: in your affiliations, please spell out USA as United States of America
- Fig 1. legend b) is missing a hyphen in TC-Python
- line 40: remove additional "s" in "should operates"
- Fig. 2
   - step 1: correct spelling of word "fucntion"
   - step 2: correct spelling of word "relavent"
   - step 3: there's a spurious black arrow across Gulliver/using
- line 53: correct spelling of word "equilibrum"
- line 78: correct rendering of alloys using subscripts as in Fig 3 caption
- line 80: I would suggest "...8.881 (in %) for Al, ..."
- line 82: I would suggest to add an article, e.g. "that the equilib_batch() feature.."
- line 85: I would suggest to add an article, e.g. "of the A380 alloy..."
- line 95: correct spelling of word "bewteen"
- Table 1, the hyperlinks of the packages don't seem to work, although the formatting in the .md file looks correct. I will check whether this behavior is known
- line 144: capital Gibbs
- line 196: capital Theriak/Domino
- line 203: please correct last author as "Martin, and SM" in paper.bib
- line 218: capital Gulliver
- line 228: capital Gibbs
- line 236: capital Schichtkristallbildung
- on the use of ChatGPT can you please clarify how it was used here in more detail?
