### 2024-07-29 #6875 @phibeck
Regarding the hyperlinks in the paper, could you add `//` after `https:`? Hopefully that fixes things, thanks
