### 2024-07-29 e587934
Revision based on editor comments
