+++ paper/paper.md
@@ -21,7 +21,7 @@ authors:
     affiliation: 1
     orcid: 0000-0001-6480-2254
 affiliations:
- - name: Materials Science and Technology Division, Oak Ridge National Laboratory, Oak Ridge, TN 37831, USA
+ - name: Materials Science and Technology Division, Oak Ridge National Laboratory, Oak Ridge, TN 37831, United States of America
    index: 1
  - name: Independent Researcher, Canada
    index: 2
@@ -40,7 +40,7 @@ In response to the need, we developed Equilipy an open-source Python package des
 # Statement of need
 Commercial thermochemical software packages such as Thermo-Calc [@ThermoCalc:2002], FactSage [@FactSage:2016], and Pandat [@Pandat:2009] are the standard for phase equilibrium calculations. However, exploring a broad spectrum of alloy chemistries is often hindered by the considerable computational time required. Such computational demands create a bottleneck in effective alloy design processes. While efforts to leverage High-Performance Computing (HPC) have been initiated to expedite computational speed, these attempts are often limited by licensing constraints such as the number of processors available for use (e.g., TC-Python permits only up to 32 logical processors per each license).
 
-Given these limitations and the pressing need for more effective alloy design processes, developing an open-source program tailored for high-throughput CALPHAD calculations becomes imperative. The envisioned program should not only eliminate the constraints on processor usage, but also ensure high calculation stability. Furthermore, the program should operates on Linux system and support parallel computing to be compatible with HPC environments. An intuitive program architecture is preferred to enhance accessibility for researchers and practitioners in the field.
+Given these limitations and the pressing need for more effective alloy design processes, developing an open-source program tailored for high-throughput CALPHAD calculations becomes imperative. The envisioned program should not only eliminate the constraints on processor usage, but also ensure high calculation stability. Furthermore, the program should operate on Linux system and support parallel computing to be compatible with HPC environments. An intuitive program architecture is preferred to enhance accessibility for researchers and practitioners in the field.
 
 # Software overview
 Equilipy is designed to compute thermodynamically stable phases from a specified *NPT* ensemble (i.e., *N*: elemental composition, *P*: pressure, *T*: temperature). The four-step process for calculating phase equilibria is illustrated in \autoref{Fig2}. This process closely mirrors experimental procedures; thus, it is highly intuitive even for the users primarily experienced in experimental work. Calculating phase equilibria involves minimizing a set of Gibbs energy functions for all relevant phases. Therefore, utilizing Equilipy necessitates a Gibbs energy database developed through the CALPHAD approach.
@@ -52,7 +52,7 @@ Equilipy is designed to compute thermodynamically stable phases from a specified
 
 Equilipy 0.1 incorporates several key features:
 
-- Phase equilibrum calculation for a single *NPT* input condition,
+- Phase equilibrium calculation for a single *NPT* input condition,
 - Parallel computations for multiple *NPT* input conditions,
 - Thermochemical property calculations,
 - Scheil-Gulliver solidification,
@@ -61,27 +61,27 @@ Equilipy 0.1 incorporates several key features:
[…]
