+++ paper/paper.bib
@@ -68,7 +68,7 @@
 }
 
 @article{Scheil:1942,
-  title={Bemerkungen zur schichtkristallbildung},
+  title={Bemerkungen zur Schichtkristallbildung},
   author={Scheil, Erich},
   journal={International Journal of Materials Research},
   volume={34},
@@ -116,7 +116,7 @@
 
 @article{MTDATA:2002,
   title={MTDATA-thermodynamic and phase equilibrium software from the national physical laboratory},
-  author={Davies, RH and Dinsdale, AT and Gisby, JA and Robinson, JAJ and Martin, andS M},
+  author={Davies, RH and Dinsdale, AT and Gisby, JA and Robinson, JAJ and Martin, and S M},
   journal={CALPHAD},
   volume={26},
   number={2},
