+++ paper/paper.bib
@@ -116,7 +116,7 @@
 
 @article{MTDATA:2002,
   title={MTDATA-thermodynamic and phase equilibrium software from the national physical laboratory},
-  author={Davies, RH and Dinsdale, AT and Gisby, JA and Robinson, JAJ and Martin, and S M},
+  author={Davies, RH and Dinsdale, AT and Gisby, JA and Robinson, JAJ and Martin, and SM},
   journal={CALPHAD},
   volume={26},
   number={2},
