### 2024-07-29 #6875 @SunyongKwon
Hi @phibeck,

Thank you for the thorough review!
Regarding the hyperlinks in Table 1, I just went through several checking in MacOS and Windows using Chrome, Safari, Microsoft Edge, and FireFox. They all work fine on my side. Do you mind letting me know a bit more in detail how you have accessed the links and failed?
