### 2024-07-29 #6875 @SunyongKwon
For other items, I have completed resolving the issues that you addressed
- [x] line 6: in your affiliations, please spell out USA as United States of America
- [x] Fig 1. legend b) is missing a hyphen in TC-Python
- [x] line 40: remove additional "s" in "should operates"
- [x] Fig. 2
    - [x] step 1: correct spelling of word "fucntion"
    - [x] step 2: correct spelling of word "relavent"
    - [x] step 3: there's a spurious black arrow across Gulliver/using
- [x] line 53: correct spelling of word "equilibrum"
- [x] line 78: correct rendering of alloys using subscripts as in Fig 3 caption
- [x] line 80: I would suggest "...8.881 (in %) for Al, ..."
To be consistent with rest of the manuscript, I changed "...8.881 at% for Al, ..." to "...8.881 mol% for Al, ..."
- [x] line 82: I would suggest to add an article, e.g. "that the equilib_batch() feature.."
- [x] line 85: I would suggest to add an article, e.g. "of the A380 alloy..."
- [x] line 95: correct spelling of word "bewteen"
- [x] line 144: capital Gibbs
- [x] line 203: please correct last author as "Martin, and SM" in paper.bib

**Issues that I could not resolve**
    The following items have rendering issue. They are properly capitalized in the `paper.bib`; however, it is automatically rendered to be small letters during compiling process. I am not sure how to correct this issue on my side.
    - [ ] line 196: capital Theriak/Domino
    - [ ] line 218: capital Gulliver
    - [ ] line 228: capital Gibbs
    - [ ] line 236: capital Schichtkristallbildung

These changes are pushed to the Github repository under joss branch.
