### 2024-07-29 #6875 @SunyongKwon
Here, I declare that the authors have not used generative pre-trained transformers (GPT) in developing the software (up to current version 0.1.4) as well as in preparing the manuscript.
