### 2024-07-29 #6875 @phibeck
> Hi @phibeck,
>
> Thank you for the thorough review! Regarding the hyperlinks in Table 1, I just went through several checking in MacOS and Windows using Chrome, Safari, Microsoft Edge, and FireFox. They all work fine on my side. Do you mind letting me know a bit more in detail how you have accessed the links and failed?

Thanks for investigating! Indeed, I don't have a problem using Chromium, only in Firefox. Issue on my side, please ignore my comment regarding this point.
