### 2024-07-29 #6875 @phibeck
@SunyongKwon thank you for incorporating my comments.

>       **Issues that I could not resolve**
>           The following items have rendering issue. They are properly capitalized in the `paper.bib`; however, it is automatically rendered to be small letters during compiling process. I am not sure how to correct this issue on my side.
>           - [ ] line 196: capital Theriak/Domino
>           - [ ] line 218: capital Gulliver
>           - [ ] line 228: capital Gibbs
>           - [ ] line 236: capital Schichtkristallbildung
You should be able to fix this by putting `{}` around strings in the .bib file.
