+++ paper/paper.bib
@@ -68,7 +68,7 @@
 }
 
 @article{Scheil:1942,
-  title={Bemerkungen zur Schichtkristallbildung},
+  title="{Bemerkungen zur Schichtkristallbildung}",
   author={Scheil, Erich},
   journal={International Journal of Materials Research},
   volume={34},
@@ -161,7 +161,7 @@
 }
 
 @article{Theriak:2010,
-  title={The computation of equilibrium assemblage diagrams with Theriak/Domino software},
+  title="{The computation of equilibrium assemblage diagrams with Theriak/Domino software}",
   author={de Capitani, Christian and Petrakakis, Konstantin},
   journal={American mineralogist},
   volume={95},
@@ -210,7 +210,7 @@
 }
 
 @article{MAGEMin:2022,
-  title={MAGEMin, an efficient Gibbs energy minimizer: application to igneous systems},
+  title="{MAGEMin, an efficient Gibbs energy minimizer: application to igneous systems}",
   author={Riel, Nicolas and Kaus, Boris JP and Green, ECR and Berlie, Nicolas},
   journal={Geochemistry, Geophysics, Geosystems},
   volume={23},
@@ -254,7 +254,7 @@
 }
 
 @article{Pelton:2017,
-  title={Scheil--Gulliver constituent diagrams},
+  title="{Scheil--Gulliver constituent diagrams}",
   author={Pelton, Arthur D and Eriksson, Gunnar and Bale, Christopher W},
   journal={Metallurgical and Materials Transactions A},
   volume={48},
