### 2024-07-29 #6875 @SunyongKwon
@phibeck  Thanks.

I corrected the references according to your suggestion.
- [x] line 196: capital Theriak/Domino
- [x] line 218: capital Gulliver
- [x] line 228: capital Gibbs
- [x] line 236: capital Schichtkristallbildung
