### 2024-07-30 #6875 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
