### 2024-07-30 #6875 @Kevin-Mattheus-Moerman
- [ ] I can see a code of conduct, but I do not see clear "contributing guidelines", if these are indeed missing, would you consider adding these and linking to them in the readme? You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

On the paper:
- [ ] Typo: `suppresing`  ->  `suppressing`
- [ ] Typo: `compatable` -> `compatible`
- [ ] This sentence sounds wrong, perhaps the words `and the publisher` were inserted the wrong way: `The United States
Government retains and the publisher, by accepting the article for publication, acknowledges that the United States Government retains a non-exclusive, paid-up, irrevocable, world-wide license to publish or reproduce the published form of this manuscript, or allow others to do so, for United States Government purposes.`. In addition, I will check with the board if we are happy with this statement. I think the statement is not necessary given "Copyright of JOSS papers is retained by submitting authors and accepted papers are subject to a [Creative Commons Attribution 4.0 International License](http://creativecommons.org/licenses/by/4.0/).". I will check with the board if we can keep the statement, but you can also decide to remove it based on what I've just shared.
