+++ CONTRIBUTING.md
@@ -1,0 +1,11 @@
+
+## Contributing
+
+Contributing to Equilipy is easy: just open a
+[pull request](https://help.github.com/articles/using-pull-requests/).
+Make `main` the destination branch on the [Equilipy
+repository](https://github.com/ORNL/Equilipy) and allow edits from
+maintainers.
+
+Your pull request must work with all current Equilipy tutorial examples
+and be reviewed by at least one Equilipy developer.
