+++ CODE_OF_CONDUCT.md
@@ -1,17 +1,5 @@
 # Contributor Covenant Code of Conduct
 
-## Contributing
-
-Contributing to Equilipy is easy: just open a
-[pull request](https://help.github.com/articles/using-pull-requests/).
-Make `main` the destination branch on the [Equilipy
-repository](https://github.com/ORNL/Equilipy) and allow edits from
-maintainers.
-
-Your pull request must work with all current Equilipy tutorial examples
-and be reviewed by at least one Equilipy developer.
-
-
 ## Our Standards
 
 We as contributors and maintainers pledge to making participation in our project and
