+++ paper/paper.md
@@ -139,7 +139,7 @@ The current version of Equilipy (0.1.4) is specifically designed for metallic sy
 
 # Acknowledgements
 
-The development of Equilipy was sponsored by the U.S. Department of Energy (DOE), Office of Energy Efficiency and Renewable Energy (EERE), Office of Sustainable Transportation, Vehicle Technologies Office (VTO). This manuscript has been authored by UT-Battelle, LLC under Contract No. DE-AC05-00OR22725 with the U.S. Department of Energy. The United States Government retains and the publisher, by accepting the article for publication, acknowledges that the United States Government retains a non-exclusive, paid-up, irrevocable, world-wide license to publish or reproduce the published form of this manuscript, or allow others to do so, for United States Government purposes. The Department of Energy will provide public access to these results of federally sponsored research in accordance with the DOE Public Access Plan (http://energy.gov/downloads/doe-public-access-plan). The authors extend their gratitude to In-Ho Jung and Jaesung Lee from the Department of Materials Science and Engineering at Seoul National University for their valuable discussions on algorithms. Special thanks are also due to Samuel T. Reeve and Gerald L. Knapp for their technical support and insightful discussion concerning programming.
+The development of Equilipy was sponsored by the U.S. Department of Energy (DOE), Office of Energy Efficiency and Renewable Energy (EERE), Office of Sustainable Transportation, Vehicle Technologies Office (VTO). This manuscript has been authored by UT-Battelle, LLC under Contract No. DE-AC05-00OR22725 with the U.S. Department of Energy. The publisher, by accepting the article for publication, acknowledges that the United States Government retains a non-exclusive, paid-up, irrevocable, world-wide license to publish or reproduce the published form of this manuscript, or allow others to do so, for United States Government purposes. The Department of Energy will provide public access to these results of federally sponsored research in accordance with the DOE Public Access Plan (http://energy.gov/downloads/doe-public-access-plan). The authors extend their gratitude to In-Ho Jung and Jaesung Lee from the Department of Materials Science and Engineering at Seoul National University for their valuable discussions on algorithms. Special thanks are also due to Samuel T. Reeve and Gerald L. Knapp for their technical support and insightful discussion concerning programming.
 
 # References
 
