### 2024-07-30 #6875 @SunyongKwon
Hi @Kevin-Mattheus-Moerman,

Thank you so much for your comments and effort to improve the quality of article. I have revised and implemented your suggestions for
- [x] Create a CONTRIBUTING.md file
- [x] Typo: suppresing -> suppressing
- [x] Typo: compatable -> compatible
- [x] 'The United States Government retains and the publisher, by accepting the article for publication, acknowledges that' -> 'The publisher, by accepting the article for publication, acknowledges that the United States Government retains'

Please note that all manuscripts published by the Oak Ridge National Laboratory require the acknowledgment and agreement. Here is an example [article](https://joss.theoj.org/papers/10.21105/joss.04115) published in JOSS with the same acknowledgment.

Thank you again for the thorough review.
