### 2024-07-30 #6875 @phibeck
Hi @Kevin-Mattheus-Moerman I also found the phrasing odd, but it seems to be official wording since it can be found all over the internet, e.g. https://www.acquisition.gov/dears/970.5227-2-rights-data-technology-transfer.
