### 2024-07-31 #6875 @SunyongKwon
Hi @Kevin-Mattheus-Moerman @phibeck,

I have update the main branch of the GitHub repository after creating `CONTRIBUTING.md`.
For consistency, i should increase the version number `0.1.4`->`0.1.5`, and provide revised Zenodo link.

@Kevin-Mattheus-Moerman
Please let me know if you have additional comment before i update the release to `0.1.5`.
