### 2024-08-02 #6875 @Kevin-Mattheus-Moerman
@SunyongKwon congratulations on this JOSS paper!

Thanks for editing @phibeck ! And a special thank you to the reviewers! @mefuller, @lanari
